Modify typing session via method, not field; rename metod onEvent to more clear; add docs; use Double instead of Flow

+++ core/analytics/api/src/main/java/com/dapsoft/wpmcounter/analytics/speed/TypingSpeedState.kt
@@ -10,7 +10,7 @@ package com.dapsoft.wpmcounter.analytics.speed
  *
  */
 sealed interface TypingSpeedState {
-    data class Active(val wordsPerMinute: Float) : TypingSpeedState
+    data class Active(val wordsPerMinute: Double) : TypingSpeedState
     object Paused : TypingSpeedState
     object Error : TypingSpeedState
 }
